% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaussianNaiveBayes.R
\name{Gaussian_Naive_Bayes}
\alias{Gaussian_Naive_Bayes}
\title{fit}
\value{
A matrix type object.
}
\description{
fit

fit

predict

predict_proba

confusion_matrix

print

summary
}
\details{
Gaussian_Naive_Bayes Class object
}
\examples{

## ------------------------------------------------
## Method `Gaussian_Naive_Bayes$fit`
## ------------------------------------------------

#NB <- Gaussian_Naive_Bayes$new()
#NB$fit(X_train, y_train)

## ------------------------------------------------
## Method `Gaussian_Naive_Bayes$predict`
## ------------------------------------------------

#y_pred <- NB$predict(Xtest)
#print(y_pred)


## ------------------------------------------------
## Method `Gaussian_Naive_Bayes$confusion_matrix`
## ------------------------------------------------

#conf_mat = NB$confusion_matrix(ytest,y_pred)
#print(conf_mat)


## ------------------------------------------------
## Method `Gaussian_Naive_Bayes$print`
## ------------------------------------------------

#print(NB)


## ------------------------------------------------
## Method `Gaussian_Naive_Bayes$summary`
## ------------------------------------------------

#NB$summary()


}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{X}}{Training set.}

\item{\code{y}}{Target variable to train the model with.}

\item{\code{params}}{Contains both mu et sd parameters.}

\item{\code{prior}}{Prior probabilities of each class in y.}

\item{\code{post}}{Post probabilities of the classed data.}

\item{\code{vars}}{Variables in X.}

\item{\code{levels_y}}{Classes in y.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Gaussian_Naive_Bayes-fit}{\code{Gaussian_Naive_Bayes$fit()}}
\item \href{#method-Gaussian_Naive_Bayes-predict}{\code{Gaussian_Naive_Bayes$predict()}}
\item \href{#method-Gaussian_Naive_Bayes-predict_proba}{\code{Gaussian_Naive_Bayes$predict_proba()}}
\item \href{#method-Gaussian_Naive_Bayes-confusion_matrix}{\code{Gaussian_Naive_Bayes$confusion_matrix()}}
\item \href{#method-Gaussian_Naive_Bayes-print}{\code{Gaussian_Naive_Bayes$print()}}
\item \href{#method-Gaussian_Naive_Bayes-summary}{\code{Gaussian_Naive_Bayes$summary()}}
\item \href{#method-Gaussian_Naive_Bayes-clone}{\code{Gaussian_Naive_Bayes$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_Naive_Bayes-fit"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_Naive_Bayes-fit}{}}}
\subsection{Method \code{fit()}}{
Train the model.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_Naive_Bayes$fit(X, y)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{a dataframe containing your training data set.}

\item{\code{y}}{a vector of type factor, character or logical containing.}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#NB <- Gaussian_Naive_Bayes$new()
#NB$fit(X_train, y_train)
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_Naive_Bayes-predict"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_Naive_Bayes-predict}{}}}
\subsection{Method \code{predict()}}{
Use the model to predict test values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_Naive_Bayes$predict(X_test, threshold = 0.001, eps = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_test}}{vector of character}

\item{\code{threshold}}{numeric value}

\item{\code{eps}}{numeric value}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
vector of character
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#y_pred <- NB$predict(Xtest)
#print(y_pred)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_Naive_Bayes-predict_proba"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_Naive_Bayes-predict_proba}{}}}
\subsection{Method \code{predict_proba()}}{
Gives the post probabilities for each row of your test set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_Naive_Bayes$predict_proba(X_test)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X_test}}{Dataframe containing your test data

@examples
#y_proba = NB$predict_proba(Xtest)
#print(y_proba)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_Naive_Bayes-confusion_matrix"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_Naive_Bayes-confusion_matrix}{}}}
\subsection{Method \code{confusion_matrix()}}{
Creates a confusion matrix confronting the classed data to the real classes.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_Naive_Bayes$confusion_matrix(y_test, y_pred)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{y_test}}{vector of characters containing test data}

\item{\code{y_pred}}{vector of characters containing predictions}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#conf_mat = NB$confusion_matrix(ytest,y_pred)
#print(conf_mat)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_Naive_Bayes-print"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_Naive_Bayes-print}{}}}
\subsection{Method \code{print()}}{
Displays information data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_Naive_Bayes$print()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#print(NB)

}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_Naive_Bayes-summary"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_Naive_Bayes-summary}{}}}
\subsection{Method \code{summary()}}{
displays information about the model and data.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_Naive_Bayes$summary()}\if{html}{\out{</div>}}
}

\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{#NB$summary()


}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Gaussian_Naive_Bayes-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Gaussian_Naive_Bayes-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Gaussian_Naive_Bayes$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
